Create index
The HTML code for the website of a gym.
The website here was designed to be simple and non fussy. The gallery images top centre shows two guys giving a high five. When attending a gym class one needs to form mini teams to go through the equipment for e.g. HIIT or TABATA classes. Mutual support and motivation increases class enjoyment an increseases health outcomes. There are also those who discuss equipment avaliable at a gym. The top left gallery image was provided to support this.