Create heap sort
Time Complexity: O(N log N)
Auxiliary Space: O(1)
Follow the given steps to solve the problem:

Build a max heap from the input data. 
At this point, the maximum element is stored at the root of the heap. Replace it with the last item of the heap followed by reducing the size of the heap by 1. Finally, heapify the root of the tree. 
Repeat step 2 while the size of the heap is greater than 1.